reorganized etherscan dtos and standardized requests and responses folders in dto, keeping entities to tables in mysql one to one
